#440 Move the fuzzy matcher to ui/common
